OpenGL Backend: Actors now have Loop event. Implemented horizontal speed and vertical speed Actors X and Y are now float variables. Fixed speed.

+++ Sabre/Trunk/OpenGL Backend/drawable.h
@@ -28,7 +28,7 @@ THE SOFTWARE.
 class pineapple::objects::Drawable
 {
     public:
-        virtual void GLDraw(int x, int y) = 0;
+        virtual void GLDraw(float x, float y) = 0;
 };
 
 #endif	/* _DRAWABLE_H */
